move frame_holder into own header w/ <<()
