+++ src/frame.cpp
@@ -11,6 +11,37 @@
 
 namespace librealsense {
 
+
+std::ostream & operator<<( std::ostream & s, const frame_interface & f )
+{
+    if( !&f )
+    {
+        s << "[null]";
+    }
+    else
+    {
+        auto composite = dynamic_cast<const composite_frame *>(&f);
+        if( composite )
+        {
+            s << "[";
+            for( int i = 0; i < composite->get_embedded_frames_count(); i++ )
+            {
+                s << *composite->get_frame( i );
+            }
+            s << "]";
+        }
+        else
+        {
+            s << "[" << f.get_stream()->get_stream_type();
+            s << "/" << f.get_stream()->get_unique_id();
+            s << " " << f.get_header();
+            s << "]";
+        }
+    }
+    return s;
+}
+
+
 std::ostream & operator<<( std::ostream & os, frame_header const & header )
 {
     os << "#" << header.frame_number;
@@ -18,6 +49,12 @@ std::ostream & operator<<( std::ostream & os, frame_header const & header )
     if( header.timestamp_domain != RS2_TIMESTAMP_DOMAIN_HARDWARE_CLOCK )
         os << "/" << rs2_timestamp_domain_to_string( header.timestamp_domain );
     return os;
+}
+
+
+std::ostream & operator<<( std::ostream & os, frame_holder const & f )
+{
+    return os << *f.frame;
 }
