+++ src/media/playback/playback_device.cpp
@@ -645,7 +645,7 @@ void playback_device::try_looping()
                 LOG_ERROR(error_msg);
                 throw invalid_value_exception(error_msg);
             }
-            LOG_DEBUG("Dispatching frame " << frame_holder_to_string(frame->frame));
+            LOG_DEBUG("Dispatching frame " << frame->frame);
 
             if (data->is<serialized_invalid_frame>())
             {
