+++ src/points.cpp
@@ -3,6 +3,7 @@
 #include "points.h"
 #include "core/video.h"
 #include "core/video-frame.h"
+#include "core/frame-holder.h"
 #include <fstream>
 
 #define MIN_DISTANCE 1e-6
