+++ src/points.h
@@ -6,8 +6,14 @@
 #include "core/extension.h"
 #include "types.h"
 
+#include <string>
+
 
 namespace librealsense {
+
+
+class frame_holder;
+
 
 class points : public frame
 {
@@ -19,5 +25,6 @@ public:
 };
 
 MAP_EXTENSION( RS2_EXTENSION_POINTS, librealsense::points );
+
 
 }  // namespace librealsense
