+++ src/rs.cpp
@@ -119,7 +119,7 @@ struct rs2_frame_queue
 {
     explicit rs2_frame_queue(int cap)
         : queue( cap, [cap]( librealsense::frame_holder const & fh ) {
-            LOG_DEBUG( "DROPPED queue (capacity= " << cap << ") frame " << frame_holder_to_string( fh ) );
+            LOG_DEBUG( "DROPPED queue (capacity= " << cap << ") frame " << fh );
         } )
     {
     }
